requirements_streamlit_cloud.txt
streamlit>=1.36.0
pyspark==3.5.3
pandas>=2.1.0
pyarrow>=14.0.0
gdown>=5.1.0
requests>=2.31.0
google-cloud-storage>=2.18.0
google-cloud-bigquery>=3.25.0
python-dotenv>=1.0.1
